+++ src/gtk4/treeview-liststore-editable/MainWindow.py
@@ -24,8 +24,8 @@ class MainWindow(Gtk.ApplicationWindow):
         super().__init__(**kwargs)
 
         self.set_title(title='Python e GTK 4: PyGObject Gtk.TreeView() editable')
-        self.set_default_size(width=1366 / 2, height=768 / 2)
-        self.set_size_request(width=1366 / 2, height=768 / 2)
+        self.set_default_size(width=int(1366 / 2), height=int(768 / 2))
+        self.set_size_request(width=int(1366 / 2), height=int(768 / 2))
 
         vbox = Gtk.Box.new(orientation=Gtk.Orientation.VERTICAL, spacing=12)
         vbox.set_homogeneous(homogeneous=True)
